Add files via upload
This is a simple front-end design for a limousine website. It lacks back-end functionalities, such as a user database to handle sign-up and log-in functions. Instead, it uses console logs for user data. For the sign-up and login functions to work properly, it needs a database, such as MongoDB, or a live server database.